+++ addons/mail/tests/__init__.py
@@ -1,5 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import test_ir_mail_server
 from . import test_link_preview
 from . import test_mail_activity
 from . import test_mail_composer
