+++ odoo/addons/base/__manifest__.py
@@ -21,7 +21,6 @@ The kernel of Odoo, needed for all installation.
         'data/report_paperformat_data.xml',
         'data/res_country_data.xml',
         'data/ir_demo_data.xml',
-        'data/ir_config_parameter_data.xml',
         'security/base_groups.xml',
         'security/base_security.xml',
         'views/base_menus.xml',
