+++ odoo/addons/base/models/ir_mail_server.py
@@ -202,24 +202,17 @@ class IrMailServer(models.Model):
 
     def _get_test_email_from(self):
         self.ensure_one()
+        email_from = False
         if from_filter_parts := [part.strip() for part in (self.from_filter or '').split(",") if part.strip()]:
-            if mail_from := next((email for email in from_filter_parts if "@" in email), None):
-                # All emails will be sent from the same address
-                return mail_from
-            # All emails will be sent from any address in the same domain
-            default_from = self.env["ir.config_parameter"].sudo().get_param("mail.default.from", "odoo")
-            if "@" not in default_from:
-                return f"{default_from}@{from_filter_parts[0]}"
-            elif self._match_from_filter(default_from, self.from_filter):
-                # the mail server is configured for a domain
-                # that match the default email address
-                return default_from
-            # the from_filter is configured for a domain different that the one
-            # of the full email configured in mail.default.from
-            return f"noreply@{from_filter_parts[0]}"
-        # Fallback to current user email if there's no from filter
-        email_from = self.env.user.email
+            # find first found complete email in filter parts
+            email_from = next((email for email in from_filter_parts if "@" in email), False)
+            # no complete email -> consider noreply
+            if not email_from:
+                email_from = f"noreply@{from_filter_parts[0]}"
         if not email_from:
+            # Fallback to current user email if there's no from filter
+            email_from = self.env.user.email
+        if not email_from or "@" not in email_from:
             raise UserError(_('Please configure an email on the current user to simulate '
                               'sending an email message via this outgoing server'))
         return email_from
@@ -531,47 +524,20 @@ class IrMailServer(models.Model):
 
     @api.model
     def _get_default_bounce_address(self):
-        '''Compute the default bounce address.
+        """ Computes the default bounce address. It is used to set the envelop
+        address if no envelop address is provided in the message.
 
-        The default bounce address is used to set the envelop address if no
-        envelop address is provided in the message.  It is formed by properly
-        joining the parameters "mail.bounce.alias" and
-        "mail.catchall.domain".
-
-        If "mail.bounce.alias" is not set it defaults to "postmaster-odoo".
-
-        If "mail.catchall.domain" is not set, return None.
-
-        '''
-        ICP = self.env['ir.config_parameter'].sudo()
-        bounce_alias = ICP.get_param('mail.bounce.alias')
-        domain = ICP.get_param('mail.catchall.domain')
-        if bounce_alias and domain:
-            return '%s@%s' % (bounce_alias, domain)
-        return
+        :return str/None: defaults to the ``--email-from`` CLI/config parameter.
+        """
+        return tools.config.get("email_from")
 
     @api.model
     def _get_default_from_address(self):
-        """Compute the default from address.
+        """ Computes the default from address. It is used for the "header from"
+        address when no other has been received.
 
-        Used for the "header from" address when no other has been received.
-
-        :return str/None:
-            If the config parameter ``mail.default.from`` contains
-            a full email address, return it.
-            Otherwise, combines config parameters ``mail.default.from`` and
-            ``mail.catchall.domain`` to generate a default sender address.
-
-            If some of those parameters is not defined, it will default to the
-            ``--email-from`` CLI/config parameter.
+        :return str/None: defaults to the ``--email-from`` CLI/config parameter.
         """
-        get_param = self.env['ir.config_parameter'].sudo().get_param
-        email_from = get_param("mail.default.from")
-        if email_from and "@" in email_from:
-            return email_from
-        domain = get_param("mail.catchall.domain")
-        if email_from and domain:
-            return "%s@%s" % (email_from, domain)
         return tools.config.get("email_from")
 
     def _prepare_email_message(self, message, smtp_session):
