+++ odoo/addons/base/tests/common.py
@@ -399,13 +399,7 @@ class MockSmtplibCase:
 
     @classmethod
     def _init_mail_gateway(cls):
-        cls.alias_bounce = 'bounce.test'
-        cls.alias_domain = 'test.mycompany.com'
-        cls.default_from = 'notifications.test'
         cls.default_from_filter = False
-        cls.env['ir.config_parameter'].sudo().set_param('mail.catchall.domain', cls.alias_domain)
-        cls.env['ir.config_parameter'].sudo().set_param('mail.default.from', cls.default_from)
-        cls.env['ir.config_parameter'].sudo().set_param('mail.bounce.alias', cls.alias_bounce)
         cls.env['ir.config_parameter'].sudo().set_param('mail.default.from_filter', cls.default_from_filter)
 
     @classmethod
